requirements.txt
pandas==2.2.2
pyarrow==17.0.0
scikit-learn==1.5.1
python-dotenv==1.0.1
SQLAlchemy==2.0.32
pika==1.3.2
textblob